build: ⬆️ update build envrionment
- update to go 1.27.1
- better devcontainer postcreate setup
- switch to oh-my-posh for devcontainer shell
- update go and alpine base image versions

+++ .devcontainer/postcreate.sh
@@ -6,25 +6,40 @@ set -x
 
 cd /workspace
 
+
+# Update JS packages with bun.
+if [[ -e ./package.json ]]; then
+    npm clean-install || exit -1
+fi
+echo 'set --export PATH "/workspace/node_modules/.bin" $PATH' >> ~/.config/fish/config.fish
+echo 'export PATH=/workspace/node_modules/.bin:${PATH}' >> ~/.bashrc
+
 # Update JS packages.
 npm clean-install || exit -1
 
 # Install Go packages.
-export PATH="$HOME/go/bin:/go/bin:/usr/local/go/bin:$PATH" && \
-    go mod tidy && \
-    go install github.com/air-verse/air@latest && \
-    go install github.com/a-h/templ/cmd/templ@latest && \
-    go install golang.org/x/tools/gopls@latest && \
-    go install github.com/sigstore/cosign/v3/cmd/cosign@latest && \
-    curl -sSfL https://golangci-lint.run/install.sh | sh -s -- -b $(go env GOPATH)/bin v2.11.4 && \
-    golangci-lint custom && \
-    mv /tmp/golangci-lint-v2 $(go env GOPATH)/bin/golangci-lint-v2
-
-# Set up fish shell.
-mkdir -p ~/.config/fish
-echo 'set --export PATH "/workspace/node_modules/.bin" $PATH' >> ~/.config/fish/config.fish
 echo 'set --export PATH "$HOME/go/bin" /go/bin /usr/local/go/bin $PATH' >> ~/.config/fish/config.fish
-echo "starship init fish | source" >>~/.config/fish/config.fish
-echo 'eval "$(starship init bash)"' >>~/.bashrc
+echo 'export PATH="$HOME/go/bin:/go/bin:/usr/local/go/bin:$PATH' >> ~/.bashrc
+if [[ -e go.mod ]]; then
+    export PATH="$HOME/go/bin:/go/bin:/usr/local/go/bin:$PATH" && \
+        go mod tidy && \
+        go install golang.org/x/tools/gopls@latest && \
+        go install github.com/air-verse/air@latest && \
+        go install github.com/a-h/templ/cmd/templ@latest && \
+        go install github.com/magefile/mage@latest && \
+        curl -sSfL https://golangci-lint.run/install.sh | sh -s -- -b $(go env GOPATH)/bin v2.13.2
+fi
+
+if [[ -e ./.custom-gcl.yaml ]]; then
+    golangci-lint custom && \
+    mv /tmp/golangci-lint-v2 $(go env GOPATH)/bin/
+fi
+
+# Install oh-my-posh.
+mkdir -p ~/.local/bin && curl -s https://ohmyposh.dev/install.sh | bash -s
+# Set up shells to use oh-my-posh.
+mkdir -p ~/.config/fish \
+    && echo "~/.local/bin/oh-my-posh init fish | source" >>~/.config/fish/config.fish \
+    && echo 'eval "$(~/.local/bin/oh-my-posh init bash)""' >>~/.bashrc
 
 exit 0
